# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: TraceLab AI API
  description: |
    REST API for TraceLab AI — intelligent quality control laboratory platform.
    
    ## Authentication
    Most endpoints require a valid JWT access token in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Access tokens expire in **15 minutes**. Use `/auth/refresh` with the refresh token to get a new one.
    
    ## Roles
    | Role | Access Level |
    |---|---|
    | `admin` | Full access to all resources and settings |
    | `lab_manager` | Manage samples, analyses, clients, products, users (except admin) |
    | `analyst` | Create/update analyses and samples; read-only on other resources |
    | `viewer` | Read-only access to all resources |
    
    ## Pagination
    List endpoints accept `page` (default: 1) and `limit` (default: 20, max: 100) query params and return a `PaginatedResponse`.
    
    ## Error Format
    All errors return a consistent JSON envelope:
    ```json
    { "statusCode": 404, "error": "Not Found", "message": "Sample not found" }
    ```
  version: 1.0.0
  contact:
    name: TraceLab AI Engineering
    email: dev@tracelab.app
servers:
  - url: http://localhost/api
    description: Local development (via NGINX proxy)
  - url: https://api.tracelab.app
    description: Production

  # AI microservice — separate server entry
  - url: http://localhost/ai
    description: AI service — local (via NGINX proxy)
  - url: https://ai.tracelab.app
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable Schemas ──────────────────────────
  schemas:
...
    UserRole:
      type: string
      enum: [admin, lab_manager, analyst, viewer]
...
    ApiError:
      type: object
      required: [statusCode, error, message]
      properties:
        statusCode: { type: integer, example: 404 }
        error:      { type: string,  example: 'Not Found' }
        message:    { type: string,  example: 'Sample not found' }

    # ── Pagination ──────────────────────────────────
    PaginationMeta:
      type: object
      properties:
        total:      { type: integer, example: 150 }
        page:       { type: integer, example: 1 }
        limit:      { type: integer, example: 20 }
        totalPages: { type: integer, example: 8 }

    # ── Auth ────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:           { type: string, format: uuid }
        name:         { type: string }
        email:        { type: string, format: email }
        role:
          $ref: '#/components/schemas/UserRole'
        laboratoryId: { type: string, format: uuid }
        isActive:     { type: boolean }
        createdAt:    { type: string, format: date-time }
        updatedAt:    { type: string, format: date-time }
    CreateUserDto:
      type: object
      required: [name, email, password, role]
      properties:
        name:     { type: string, example: 'Maria Costa' }
        email:    { type: string, format: email }
        password: { type: string, minLength: 8 }
        role:
          $ref: '#/components/schemas/UserRole'
    UpdateUserDto:
      type: object
      properties:
        name:     { type: string }
        role:
          $ref: '#/components/schemas/UserRole'
        isActive: { type: boolean }

    # ── Laboratory ──────────────────────────────────
...
    Product:
      type: object
      properties:
        id:           { type: string, format: uuid }
        laboratoryId: { type: string, format: uuid }
        name:         { type: string }
        sku:          { type: string }
        barcode:      { type: string }
        qrcode:       { type: string }
        category:     { type: string }
        aiTags:       { type: object }
        images:       { type: array, items: { type: string } }
        createdAt:    { type: string, format: date-time }
    CreateProductDto:
      type: object
      required: [name]
      properties:
        name:     { type: string }
        sku:      { type: string }
        barcode:  { type: string }
        qrcode:   { type: string }
        category: { type: string }
    ScanRequest:
      type: object
      required: [code]
      properties:
        code: { type: string, example: '7891234567890', description: 'QR code or barcode value' }

    # ── Sample ──────────────────────────────────────
...
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }

  # ─── Reusable Responses ─────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Insufficient role permissions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }

# ─── Global Security ────────────────────────────
security:
  - bearerAuth: []

# ═══════════════════════════════════════════════
# PATHS
# ═══════════════════════════════════════════════
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users in the laboratory
      x-roles: [admin, lab_manager]
      parameters:
        - { $ref: '#/components/parameters/PageParam' }
        - { $ref: '#/components/parameters/LimitParam' }
        - name: role
          in: query
          schema: { $ref: '#/components/schemas/UserRole' }
        - name: isActive
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationMeta'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Users]
      summary: Create a new user
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserDto' }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [Users]
      summary: Get user by ID
      x-roles: [admin, lab_manager]
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Users]
      summary: Update user
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserDto' }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Users]
      summary: Deactivate user (soft delete)
      x-roles: [admin]
      responses:
        '204': { description: User deactivated }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─── LABORATORIES ────────────────────────────────
...
  /products:
    get:
      tags: [Products]
      summary: List products
      parameters:
        - { $ref: '#/components/parameters/PageParam' }
        - { $ref: '#/components/parameters/LimitParam' }
        - name: search
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated list of products
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationMeta'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Product' }

    post:
      tags: [Products]
      summary: Create a product
      x-roles: [admin, lab_manager]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProductDto' }
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
  /products/scan:
    post:
      tags: [Products]
      summary: Look up product by QR code or barcode
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScanRequest' }
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Products]
      summary: Get product by ID
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Products]
      summary: Update product
      x-roles: [admin, lab_manager]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProductDto' }
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
    delete:
      tags: [Products]
      summary: Delete product
      x-roles: [admin]
      responses:
        '204': { description: Product deleted }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─── SAMPLES ─────────────────────────────────────
...